S4 class for promoter annotation data for a specific annotation version

S4 class for promoter annotation data for a specific annotation version

PromoterAnnotation(
  intronRanges = GRanges(),
  promoterIdMapping = data.frame(),
  promoterCoordinates = GRanges()
)

intronRanges(x)

# S4 method for PromoterAnnotation
intronRanges(x)

promoterIdMapping(x)

# S4 method for PromoterAnnotation
promoterIdMapping(x)

promoterCoordinates(x)

# S4 method for PromoterAnnotation
promoterCoordinates(x)

intronRanges(x) <- value

# S4 method for PromoterAnnotation
intronRanges(x) <- value

promoterIdMapping(x) <- value

# S4 method for PromoterAnnotation
promoterIdMapping(x) <- value

promoterCoordinates(x) <- value

# S4 method for PromoterAnnotation
promoterCoordinates(x) <- value

Arguments

intronRanges
A GRanges object containing annotated intron ranges

promoterIdMapping
A data.frame containing mapping between transcript, TSS, promoter and gene ids

promoterCoordinates
A GRanges object containing promoter coordinates

x
A PromoterAnnotation object

value
intronRanges, promoterIdMapping or promoterCoordinates to be assigned

Value

A promoter annotation object with three slots: intronRanges, promoterIdMapping and promoter Coordinates

Functions

intronRanges: Getter for intronRanges

intronRanges,PromoterAnnotation-method: Getter for intronRanges

promoterIdMapping: Getter for promoterIdMapping

promoterIdMapping,PromoterAnnotation-method: Getter for promoterIdMapping

promoterCoordinates: Getter for promoterCoordinates

promoterCoordinates,PromoterAnnotation-method: Getter for promoterCoordinates

intronRanges<-: Setter for intronRanges

intronRanges<-,PromoterAnnotation-method: Setter for intronRanges

promoterIdMapping<-: Setter for promoterIdMapping

promoterIdMapping<-,PromoterAnnotation-method: Setter for promoterIdMapping

promoterCoordinates<-: Setter for promoterCoordinates

promoterCoordinates<-,PromoterAnnotation-method: Setter for promoterCoordinates

Slots

intronRanges

A GRanges object. The intron ranges annotated with the promoter information.

promoterIdMapping

A data.frame object. The id mapping between transcript ids, names, TSS ids, promoter ids and gene ids.

promoterCoordinates

A GRanges object. Promoter coordinates (TSS) with gene id and internal promoter state

Examples

promoterAnnotation <- PromoterAnnotation()
intronRanges(promoterAnnotation) <- intronRanges(
                                            promoterAnnotation.gencode.v19)
promoterIdMapping(promoterAnnotation) <- promoterIdMapping(
                                            promoterAnnotation.gencode.v19)
promoterCoordinates(promoterAnnotation) <- promoterCoordinates(
                                            promoterAnnotation.gencode.v19)