Calculate the promoter read counts using junction read counts approach for all the input junction files

Calculate the promoter read counts using junction read counts approach for all the input junction files

calculatePromoterReadCounts(
  promoterAnnotation,
  files = NULL,
  fileLabels = NULL,
  fileType = NULL,
  genome = NULL,
  numberOfCores = 1
)

Arguments

promoterAnnotation
A PromoterAnnotation object containing the reduced exon ranges, annotated intron ranges, promoter coordinates and the promoter id mapping

files
A character vector. The list of junction or BAM files for which the junction read counts will be calculated

fileLabels
A character vector. The labels of junction or BAM files for which the junction read counts will be calculated. These labels will be used as column names for the output data.frame object

fileType
A character. Type of the junction bed or bam file, either 'tophat', 'star' or 'bam

genome
A character. Genome version used. Must be specified if input is a BAM file. Defaults to NULL

numberOfCores
A numeric value. The number of cores to be used for counting junction reads. Defaults to 1 (no parallelization). This parameter will be used as an argument to BiocParallel::bplapply

Value

A data.frame object. The number of junction reads per promoter (rows) for each sample (cols)

Examples

files <- list.files(system.file('extdata/testdata/tophat2',
                    package = 'proActiv'),
                    full.names = TRUE, pattern = 'sample')
fileLabels <- c('sample1', 'sample2')
promoterAnnotation <- promoterAnnotation.gencode.v19
promoterReadCounts <- calculatePromoterReadCounts(promoterAnnotation,
                                                   files,
                                                   fileLabels,
                                                   fileType = 'tophat',
                                                   genome = NULL,
                                                   numberOfCores = 1)

#> [1] "Processing: C:/Users/josep/AppData/Local/Temp/Rtmp6vqAvQ/temp_libpath2904f187c3e/proActiv/extdata/testdata/tophat2/sample1.bed"
#> [1] "File loaded into memory"
#> [1] "Calculating junction counts"
#> [1] "Processing: C:/Users/josep/AppData/Local/Temp/Rtmp6vqAvQ/temp_libpath2904f187c3e/proActiv/extdata/testdata/tophat2/sample2.bed"
#> [1] "File loaded into memory"
#> [1] "Calculating junction counts"