Visualizes heatmap of features for samples

Visualizes heatmap of features for samples

plotHeatmap(
  result,
  by = "absolutePromoterActivity",
  features = NULL,
  cex.legend = 0.75,
  cex.row = NULL,
  cex.col = NULL,
  row.margin = 5,
  col.margin = 12,
  col = NULL,
  breaks = NULL,
  palette = "bluered"
)

Arguments

result
A SummarizedExperiment object return by proActiv, with assays giving promoter counts and activity with gene expression stored as metadata. rowData contains promoter metadata and absolute promoter activity summarized across conditions. Condition must be provided.

by
A character vector. The assay to visualize the heatmap for. One of promoterCounts, normalizedPromoterCounts, absolutePromoterActivity and geneExpression (unambiguous substrings can be supplied). Defaults to absolutePromoterActivity.

features
Features to visualize. Either a list of promoterIds or geneIds. The features must correspond to the assay is used, i.e., if promoter assays are used, features must be promoterIds, while if gene expression assay is used, features must be geneIds. Defaults to NULL (visualizes all features of the assay).

cex.legend
A numeric value. Legend size.

cex.row
A numeric value. Row label size.

cex.col
A numeric value. Column label size.

row.margin
A numeric value. Row margins.

col.margin
A numeric value. Column margins.

col
A vector of colours. Length should correspond to number of experimental conditions. Defaults to NULL.

breaks
A numeric vector. Breaks for heatmap plotting.

palette
A character vector. One of bluered, redblue, redgreen, greenred. Defaults to bluered.

Value

Displays heatmap.

Examples

files <- list.files(system.file('extdata/vignette/junctions', 
                       package = 'proActiv'), 
                       full.names = TRUE, pattern = 'replicate5')
promoterAnnotation <- promoterAnnotation.gencode.v34.subset
result <- proActiv(files = files,
                       promoterAnnotation  = promoterAnnotation,
                       condition = rep(c('A549', 'HepG2'), each=1),
                       fileLabels = NULL,
                       ncores = 1)
result <- result[complete.cases(assays(result)[[1]]),] 
plotHeatmap(result)