Performs principal component analysis

Performs principal component analysis

plotPCA(
  result,
  by = "absolutePromoterActivity",
  main = NULL,
  col = NULL,
  alpha = 0.75,
  cex.size = 2
)

Arguments

result
A SummarizedExperiment object return by proActiv, with assays giving promoter counts and activity with gene expression stored as metadata. rowData contains promoter metadata and absolute promoter activity summarized across conditions. Condition must be provided.

by
A character vector. The assay to perform principal component analysis by. One of promoterCounts, normalizedPromoterCounts, absolutePromoterActivity and geneExpression (unambiguous substrings can be supplied). Defaults to absolutePromoterActivity.

main
A character vector. Plot title (optional). Defaults to NULL.

col
A vector of colours. If NULL, uses standard ggplot colours. Defaults to NULL.

alpha
A numeric value in between 0 and 1. Determines point transparency.

cex.size
A numeric value. Determines point size.

Value

PCA plot.

Examples

files <- list.files(system.file('extdata/vignette/junctions', 
                       package = 'proActiv'), 
                       full.names = TRUE, pattern = 'replicate5')
promoterAnnotation <- promoterAnnotation.gencode.v34.subset
result <- proActiv(files = files,
                       promoterAnnotation  = promoterAnnotation,
                       condition = rep(c('A549', 'HepG2'), each=1),
                       fileLabels = NULL,
                       ncores = 1)
result <- result[complete.cases(assays(result)[[1]]),]
plotPCA(result)