Prepare promoter annotation data for the user specified txdb object

Prepare promoter annotation data for the user specified txdb object

preparePromoterAnnotation(txdb, file, species)

Arguments

txdb
A txdb object. The txdb of the annotation version for which promoters will be identified. Either `txdb` or `file` argument must be specified, but not both.

file
A character object. The file path to a gtf/gff or txdb of the annotation version for which promoters will be identified. Either `txdb` or `file` argument must be specified, but not both.

species
A character object. The genus and species of the organism to be used in keepStandardChromosomes(). Supported species can be seen with names(genomeStyles()).

Value

A PromoterAnnotation object. The annotated intron ranges, promoter coordinates and the promoter id mapping are attributes of the promoter annotation data.

Examples

txdbPath <- system.file('extdata/vignette/annotations/',
                            'gencode.v34.annotation.chr22.sqlite',
                             package = 'proActiv')
txdb <- AnnotationDbi::loadDb(txdbPath)

#> Loading required package: GenomicFeatures

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: 'BiocGenerics'

#> The following objects are masked from 'package:parallel':
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from 'package:base':
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which, which.max, which.min

#> Loading required package: S4Vectors

#> Loading required package: stats4

#> 
#> Attaching package: 'S4Vectors'

#> The following object is masked from 'package:base':
#> 
#>     expand.grid

#> Loading required package: IRanges

#> 
#> Attaching package: 'IRanges'

#> The following object is masked from 'package:grDevices':
#> 
#>     windows

#> Loading required package: GenomeInfoDb

#> Loading required package: GenomicRanges

#> Loading required package: AnnotationDbi

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

promoterAnnotation <- preparePromoterAnnotation(txdb = txdb,
                                                species = 'Homo_sapiens')

#> [1] "Extract exons by transcripts..."
#> [1] "Identify overlapping first exons for each gene..."
#> [1] "Prepare mapping between transcripts, tss, promoters and genes..."
#> [1] "Prepare annotated intron ranges..."
#> [1] "Annotating reduced exon ranges..."
#> [1] "Prepare promoter coordinates and first exon ranges..."