Wrapper function returning Summarized Experiment object giving promoter counts and activity

Wrapper function returning Summarized Experiment object giving promoter counts and activity

proActiv(
  files,
  promoterAnnotation,
  fileLabels = NULL,
  condition = NULL,
  genome = NULL,
  ncores = 1
)

Arguments

files
A character vector. The list of input files for which the junction read counts will be calculated

promoterAnnotation
A PromoterAnnotation object containing the intron ranges, promoter coordinates and the promoter id mapping

fileLabels
A character vector. The labels of input files for which the junction read counts will be calculated. These labels will be used as column names for each output data.frame object. If not provided, filenames will be used as labels. Defaults to NULL

condition
A character vector. The condition to which each sample belong to. Must correspond to the order of the files. If supplied, results are summarized by condition. Defaults to NULL

genome
A character. Genome version. Must be specified if input file type is a BAM file. Defaults to NULL

ncores
A numeric value. The number of cores to be used for counting junction reads. Defaults to 1 (no parallelization). This parameter will be used as an argument to BiocParallel::bplapply

Value

A SummarizedExperiment object with assays giving promoter counts and activity with gene expression stored as metadata. rowData contains promoter metadata and absolute promoter activity summarized across conditions (if condition is provided)

Examples

files <- list.files(system.file('extdata/vignette/junctions',
                       package = 'proActiv'),
                       full.names = TRUE, pattern = 'replicate5')
promoterAnnotation <- promoterAnnotation.gencode.v34.subset
result <- proActiv(files = files,
                       promoterAnnotation  = promoterAnnotation,
                       fileLabels = NULL,
                       ncores = 1)

#> [1] "Processing: C:/Users/josep/OneDrive/Documents/R/win-library/4.0/proActiv/extdata/vignette/junctions/SGNEx_A549_Illumina_replicate5-run1.subset.SJ.out.tab.gz"
#> [1] "File loaded into memory"
#> [1] "Calculating junction counts"
#> [1] "Processing: C:/Users/josep/OneDrive/Documents/R/win-library/4.0/proActiv/extdata/vignette/junctions/SGNEx_HepG2_Illumina_replicate5-run1.subset.SJ.out.tab.gz"
#> [1] "File loaded into memory"
#> [1] "Calculating junction counts"
#> [1] "Calculating normalized read counts..."

#> converting counts to integer mode

#> [1] "Calculating log2 absolute promoter activity..."
#> [1] "Calculating gene expression..."
#> [1] "Calculating relative promoter activity..."
#> [1] "Calculating positions of promoters..."