Prepares promoter annotation from a gtf or txdb

Prepares promoter annotation from a gtf or txdb

preparePromoterAnnotation(txdb, file, species)

Arguments

txdb
A txdb object. The txdb of the annotation version for which promoters will be identified. Either `txdb` or `file` argument must be specified, but not both.

file
A character object. The file path to a gtf/gff or txdb of the annotation version for which promoters will be identified. Either `txdb` or `file` argument must be specified, but not both.

species
A character object. The genus and species of the organism to be used in keepStandardChromosomes(). Supported species can be seen with names(genomeStyles()).

Value

A PromoterAnnotation object. The annotated intron ranges, promoter coordinates and the promoter id mapping are attributes of the promoter annotation data.

Examples

txdbPath <- system.file('extdata/vignette/annotations/',
                            'gencode.v34.annotation.subset.sqlite',
                             package = 'proActiv')
txdb <- AnnotationDbi::loadDb(txdbPath)
promoterAnnotation <- preparePromoterAnnotation(txdb = txdb,
                                                species = 'Homo_sapiens')